ilb/metrics: drop references to cilium-cli code
Currently, the loadbalancer-metrics operator cell has a dependency on
the cilium-cli code, which is problematic as it gets removed in stable
branches. Let's rework this logic to make use of the same clientset
leveraged by the rest of the operator modules, and drop the dependency.
To keep things simple for the moment, a stripped down version of the
`ExecInPod` logic is re-implemented locally. Longer term, it could be
worth extracting it into a common package, to prevent duplication.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>