hubble: automatically infer Cilium builder image in Makefiles
Currently, these Makefiles hard-code the Cilium builder image version,
expecting it to be updated by the update-cilium-builder-image.sh script.
However, this is not the case, as the script only updates the Dockerfiles.
Rather than modifying the script, let's just retry the builder image
dynamically from the Cilium agent Dockerfile, to ensure that we always
use the most up-to-date version.

Additionally re-run `make -C enterprise enterprise-proto` and committed
the resulting changes.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>