egressgwha: Add next hop for egw IPAM routes
Add a next hop for each route installed for egress gateway IPAM feature.
The next hop to be used is fetched from the default route in the main
routing table for the same egress interface.

Before this change, the connectivity to IPs not directly connected with
the egress gateway node (e.g: a public IP like 8.8.8.8) was broken from
pods running on non-gateway nodes. That happened because the routes
installed in the egress gateway IPAM table, needed for the datapath FIB
lookup that redirect the packet to the proper egress interface, were
matched and no fallback to the main routing table was happening.

On a non-gateway node:

```
root@kind-control-plane:/# ip route get fibmatch 8.8.8.8 from 172.18.0.64 oif eth0
8.8.8.8 dev eth0 table 2050 proto kernel scope link src 172.18.0.64
```

The entry in table 2050 (egw-ipam table) is matched and the next hop is
missing, while on the gateway node:

```
root@kind-control-plane:/# ip route get fibmatch 8.8.8.8 from 172.18.0.2 oif eth0
default via 172.18.0.1 dev eth0
```

the main routing table entry is matched and the next hop is found.  The
commit reflects the next hop in the main routing table to table 2050 in
order to have connectivity to public IPs.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>