bgpv2: Refactor NodeConfig and ClusterConfig status reconciliation
In the upcoming Route Reflector implementation, we need to figure out
the in-cluster BGP peering topology from multiple ClusterConfigs.
However, we are currently mixing up the logic of such a "global"
reconciliation (NodeConfig update/delete, nodeSelector conflict
detection) and ClusterConfig status ("local") reconciliation. It makes
hard to implement Route Reflector logic on top of the current
implementation.

This commit decouples "global" reconciliation and ClusterConfig-specific
reconciliation. In the new implementation, all the "global" operations
are performed first and ClusterConfig status reconciliation comes later.

In the current implementation, we List() k8s objects like CiliumNode or
NodeConfig multiple times from the store. This is wasteful and also may
cause some inconsistency between different parts of the reconciliation
because for each List(), the objects may be updated.

In the new implementation, we List() relevant k8s objects once at the
beginning and never List() or GetByName() them again in the later
reconciliation. All List()'ed objects are cached in the `reconcileCache`
data structure. We also do some indexing (e.g. creating a map of name =>
object) whenever it speeds up later reconciliation logic.

As a side-effect of this refactoring, now we have a full view of the
nodeSelector conflict. So, we can now calculate the conflict
relationship even before we create the NodeConfig. Therefore we can get
rid of the current conflict detection mechanism (checking the owner of
the NodeConfig) which can only report the partial conflict relationship.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>