bgpv2: Introduce Enterprise Neighbor Reconciler
The upcoming Route Reflector support introduces enterprise-specific
logic for neighbor reconciler. As a preparation for it, introduce
Enterprise Neighbor Reconciler. Similar to the Enterprise Service
Reconciler, it replace the OSS neighbor reconciler.

The current code is almost identical the OSS code. The only difference
is it reconciles against enterprise resource with upgrader mechanism.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>