gha/panda: tolerate minor differences in desired workers count
Allow for a 2 per thousand toleration in the expected nodes count, to
continue with the test even if a tiny fraction of the workers failed
bootstrapping and joining the cluster. This toleration takes effect
only for clusters with more than 500 nodes (which are more likely to
encounter minor problems), as it gets rounded to 0 otherwise.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>