egressgwha: Keep healthy nodes across restart
After an operator restart, we do not have any information about the
health status of all the nodes previously selected as gateways.  In
order to collect each node health status, a health checker goroutine
that probes the node is started, but until we receive the first health
probe verdict, we consider the node as unhealthy. As such, the node is
considered not eligible to be an active gateway.

As a consequence, all the status stored in the IEGP received from the
apiserver, is blanked after a restart. Therefore, a spurious
reconciliation on the gateway nodes is triggered, leading to unwanted
rewrite of the egress gateway ha BPF map, a shuffle of the previous IPAM
addresses allocation and so on. Ultimately this leads to possible
traffic disruption for the pods selected by the policies.

To avoid this, we initially consider all the previous healthy gateways
as healthy after the operator restart.  Their status is properly updated
after the health checker receives the verdict of the first probe.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>