k8s: Promote core enterprise BGP CRDs to v1 version
Introduces the v1 version for IsovalentBGPClusterConfig,
IsovalentBGPPeerConfig, IsovalentBGPAdvertisement,
IsovalentBGPNodeConfig and IsovalentBGPNodeConfigOverride CRDs
and makes this version the storage version.

As part of the promotion, the deprecated field in IsovalentBGPPeerConfig
`spec.transport.localPort` has been dropped, and the unused fields
Group and Kind in the PeerConfigReference struct used in the
IsovalentBGPClusterConfig and IsovalentBGPNodeConfig have been dropped.

The IsovalentBGPPeerConfigSpec is now defined inline instead of
embedding the OSS CiliumBGPPeerConfigSpec to allow the localPort
deprecation mentioned above.

Other than that, the CRD definitions are copied-over to v1 without any
other changes, to ensure the automatic version conversion by the apiserver
works fine. The content of the dropped deprecated fields will be ignored
during the automatic migration.

Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>