operator: Use v1 version of enterprise BGP APIs in the operator
Change use of v1alpha1 version of recently promoted enterprise
BGP CRDs to the v1 version in the operator codebase.

The only non-trivial change is in the mapping logic in
createOSSPeerConfig() where we now need to explicitly map
individual fields if IsovalentBGPPeerConfigSpec to
CiliumBGPPeerConfigSpec because of the CRD changes related
to the LocalPort deprecation.

Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>