egressgwha: Fix allocations with unbalanced affinity zones
The IPAM allocation algorithm had a bug in case of unbalanced affinity
zones, that is, zones with different number of active gateways.
Specifically, when selecting the next zones to consider for the next
allocation round, the algorithm was prioritizing the ones with fewer
allocations, but was still considering zones with no more pending
requests.

The commit fixes the issue and expand the unit testing to cover the
unbalanced zones scenario.

Fixes: 9620d082a6 ("egressgwha: Improve IPAM allocation strategy with AZ affinity")
Reported-by: Yusuke Suzuki <yusuke.suzuki@isovalent.com>
Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>