ilb: validate t1-only deploymentMode with hostname based backends
This commit introduces additional validation for the cases where
a LBService with forceDeploymentMode `t1-only` uses LBBackendPools
with backendType `Hostname`.

These cases are validated and incompatibilities are reported via the
status conditions on the LBService resource.

```
status:
  addresses: {}
  applications:
    tcpProxy:
      deploymentMode: t1-only
  conditions:
  ...
  - lastTransitionTime: "2025-02-17T09:11:07Z"
    message: forceDeploymentMode t1-only is incompatible with LBBackendPools of type
      Hostname [lb-2]
```

Note: Unfortunately it's' not possible to validate this via K8s CEL (as we do
      for the other cases, as the informations are spread over two CRDs)

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>