envoy: remove duplicated service/endpointslice informers
For cases when envoyConfig was enabled, we were starting separate
informers for services and endpoints. This resulted in a duplicated
informers, adding overhead for both cilium-agent and k8s apiserver.
Before, existing informer did not handle headless services, which was
changed in #35838 so now we can rely on a single informer instead.

Signed-off-by: Marcel Zieba <marcel.zieba@isovalent.com>