ilb: add support to filter ILB E2E tests by name
With the migration from the ILB E2E tests from `go test` to
Cilium CLI, the possibility to filter tests by name does no
longer exist.

Therefore, this commit is adding a new flag `--run` to the CLI
command that is mimicing the same behaviour. This is very handy
during development.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>