enterprise: Add kubevirt migration logic
This commit adds a new operator cell, responsible for supporting the
migration of KubeVirt VM pods if they are running in bridge networking
mode and have the "kubevirt.io/allow-pod-bridge-network-live-migration"
annotation.

Migration works by relying on the `ipmigration` package, which allows
the target pod to become ready in detached state (and thus allowing
KubeVirt to transfer VM state) without having to perform an IPAM IP
request for the IP still in use by the source VM pod. The main job of
therefore is to detach/attach KubeVirt VM pods based on their migration
state. This works as follows:

All pods belonging to the same KubeVirt VM (as indicated by the
"vm.kubevirt.io/name" pod label) are grouped together. This cell then
determines which of these pods is considered the primary pod, i.e. the
pod currently designated to receive the VM's traffic. The primary pod is
determined based on two pod properties:

 1. The primary pod must have a "kubevirt.io/nodeName" label whose value
    matches the name of the K8s node the pod is currently running on.
    This is the main indicator from KubeVirt that the pod is the one which
    should receive traffic.
 2. Any pod which looks like to be the new primary must have a
    "kubevirt.io/migration-target-ready-timestamp" annotation which
    contains a more recent timestamp than the previous primary. This ensures
    that we are not flapping between the new and old primary while the
    "kubevirt.io/nodeName" label is updated on both pods.

Once the set of pods and the primary is calculated, the cell then sets
the "cni.v1alpha1.isovalent.com/detached" on all non-primary pods, and
ensures the primary pod is attached by removing the same annotation on
the primary pod if present. This way, we ensure that the primary pod is
the only attached pod of the set of pods belonging to a KubeVirt VM and
thus is the only pod able to receive traffic. During a migration, this
means that the target pod starts in detached mode, once it becomes the
primary however we attach it and detach the source pod.