ilb: report applied deployment mode for UDPProxy
Currently, we're missing the reported status, which deployment mode
actually got applied / automatically evaluated for the UDPProxy.
(This is especially useful for `forceDeploymentMode=auto` (default)
where the control plane automatically detects the best mode.

Therefore, this commit is introducing the respective fields on the
status of the CRD `LBService` (analog to TCPProxy) and updates them
during the reconciliation.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>