ilb: LBVIP: fix status condition if no IP is assigned
Currently, the status of the CRD `LBVIP` isn't properly reflecting
the condition `lb.cilium.io/IPv4AddressAllocated` if no external IP
is assigned to the underlying K8s Service.

It's only checking for status conditions but doesn't check whether
an IPv4 is set at all.

This is the case if IPAM isn't configured correctly (e.g. no
`CiliumLoadbalancerIpPool` is created at all)

Therefore, this commit is adjusting this behaviour by introducing
a new status condition value `AddressNotAllocated` that is set
if no IPv4 is allocated.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>