ilb: fix wrong proxyprotocol test
This commit fixes two ILB proxyprotocol related tests that are
testing disallowed proxyprotocol versions. These only succeed
because the validation just adds the info about the failed
validation into the status condition of the LBService but
continues with the reconciliation.

With the change of aborting the LBService reconciliation in case
of any validation issues - these tests no longer pass.

Therefore, this commit changes the assertion of these
tests.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>