ilb: fix hostname-based backend validation
Currently, the hostname-backend & t1-only validation might fail
if not all backends are indexed yet.

```
2025-02-20T12:59:14.744386928Z time=2025-02-20T12:59:14Z level=error msg="Observed a panic" module=enterprise-operator.operator.operator-controlplane.leader-lifecycle.controller-runtime controller=lbservice controllerGroup=isovalent.com controllerKind=LBService LBService.name=udp-proxy-connectionfiltering--t1-only-allow-by-sourceip LBService.namespace=default namespace=default name=udp-proxy-connectionfiltering--t1-only-allow-by-sourceip reconcileID=260759c9-aebe-4c1c-b34b-44931ebdef76 panic="runtime error: invalid memory address or nil pointer dereference"..."
```

This commit is fixing this by skipping the backends that aren't
available yet.

Note: The existence of the backends is checked by another validation.
And as soon as the backends gets indexed - a new reconciliation
gets triggerd too.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>