.github: Use dedicated go cache on CI
[ upstream commit 79653ef64ddd8429b9eff9b50aa305c44c3803b0 ]

Commit 8f02631f230c changed the builder.sh script to reuse host go build
caches by default in order to optimize the build time for local
environments. However, in some situations on GitHub infrastructure the
default path is not writable for the actions workflow, which results in
error messages reported like this:

  ... mkdir /home/runner/go/pkg/mod/cache: permission denied

Fix this by defining a go build cache directory under /tmp, like with
the other GitHub workflows that use go build caches.

Fixes: 8f02631f230c ("contrib: Reuse local go cache in builder")
Reported-by: Tam Mach <tam.mach@cilium.io>
Signed-off-by: Joe Stringer <joe@cilium.io>
Signed-off-by: Joe Stringer <joe@isovalent.com>