fqdnha/relay: stream FQDN selectors to proxy
This introduces a new (enterprise-only) statedb table which stores FQDN
selectors to be forwarded to the proxy. Using a table and a change
iterator is convenient in that it automatically implements the
ListAndWatch semantics necessary - a newly connected proxy will get a
full view of what FQDNSelectors are around. The bookmark update type
also makes it clear when the initial List has completed.

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>