egressgwha: use ifindex of egress interface to avoid FIB lookup
Add a new version of the EGWHA policy map, with enough space to store
the ifindex of the selected egress interface. We continue to populate
both versions of the map, so that on downgrade everything is still in
place for a datapath that doesn't support the v2 map.

In the datapath, return the ifindex on a EGWHA policy lookup. There is
dormant code in the caller that will magically use this information to
select the egress interface, without any FIB lookup and custom route
installation.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>