contrib: Duplicate enterprise-kind-loadbalancer.yaml
This commit creates a copy of enterprise-kind-loadbalancer.yaml.

The non-modified copy is used for development purpose when installing
WIP ILB on Kind (nothing changed).

The original enterprise-kind-loadbalancer.yaml was extended to point
the images to the latest built by CI from the "main-ce" branch. This
allows users to try ILB on Kind w/o requiring them to build it first.

Signed-off-by: Martynas Pumputis <martynas@isovalent.com>