ent/egressgw: optimize logging in updateEgressRules
Constructing the extended logger object is potentially expensive in
terms of both CPU time and allocations (then reflected in the GC
process). For instance, logging related operations can account for a
third of the total `updateEgressRules` CPU time, even if debug logs
are disabled. In this specific context, the biggest offender is the
`joinStringers` operation, especially when a significant number of
gateways is matched by a given policy. Yet, in normal conditions,
the logger object is then never expected to be used. Hence, let's
simply reorganize a bit the structure to avoid initializing it
whenever not necessary, at the cost of slight code duplication.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>