ilb: use annotation loadbalancer.isovalent.com/node=t1 for metrics filter
Currently, the ILB metrics logic uses the label `service.cilium.io/node`
to filter for the ILB related T1 services.

To reduce the number of "dependencies" to this label, this commit is introducing
a new annotation `loadbalancer.isovalent.com/type=t1` on the generated T1 service
that can be used to check whether a K8s Service is generated by ILB and
used as T1 service.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>