.github: Add mindfulness check
This check looks for a diff in the files that are intended to be
unmodified from the upstream codebase, and fails out if those files have
not been added to an exclusion list within the newly added script.

Notably on main-ce as of now it appears to highlight several changes
that are innocuous upstream which we have failed to adopt during a merge
of the upstream tree into the internal tree. These appear likely to
diverge from upstream due to improper conflict resolution.

Signed-off-by: Joe Stringer <joe@isovalent.com>