ilb: enable T2 UDP accesslogging by default
This commit is enabling T2 UPD proxy accesslogging by default.

In addition it changes the description of all UDP related flags
to clarifiy tat it's about the UDP proxy and NOT the UDP listener
access logging (enable & format).

That's also the reason why TCP accesslogging is still disabled by
default. There it's the accesslog of the TCP listener - which is
shared by healthchecking, HTTP & TCPProxy.

Accesslogging for the TCP proxy is enabled by default.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>