ilb: introduce LbTestFunc.RunTestCase to implicitly cleanup between testcases
Currently, ILB tests with singlenode mode are failing for
tests that use multiple testcases. The rason is that
re-applying BGP config fails because it already exists - and
the cleanup (that would remove existing test-resources isn't executed).

Therefore, this commit introduces a new helper function `RunTestCase`
on `LbTestFunc` that expects a test-case "function". In addition to
execute the function it also executes the cleanup after each testcase
execution. This way singlenode mode should work again.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>