.github: Update Ubuntu 24.04 apt repositories
Revert "Revert "integration: Bump ubuntu to 24.04 for arm runners"".

This partially reverts commit 751aaff1dae7cf8ab8957403869cf4d050441eb7.

This commit restores all of the generic changes to the workflow that do
not actually apply the changeover to ubuntu-24.04-arm64. This reduces
the diff between main-ce and main to a minimal amount that is tracked by
an issue to resolve it: https://github.com/isovalent/cilium/issues/6375

Signed-off-by: Joe Stringer <joe@isovalent.com>