cli/lb: Fix BGP Peers reporting
In the BGP Peers status we want to report whether for a given svc all peers
received BGP routes from all T1 nodes (or the ones to which the peers
have established BGP sessions).

Previously, we changed the activePeers state when iterating over the BGP
routes for a given service. This became problematic when there were more
than one T1 node. For example, with one peer and two T1 nodes, we
reported "DEG [1/2]" instead of OK [2/2]" because of the aforementioned
state change.

Signed-off-by: Martynas Pumputis <martynas@isovalent.com>