TESTOWNERS: Override ownership for internal teams
The CODEOWNERS file only really handles reviews for enterprise code,
then leaves the review for OSS changes up to core-structure team. The
goal for this is to encourage good maintenance practices for managing
modularity and minimizing divergence from the upstream OSS tree, thereby
allowing us to benefit from the latest and greatest OSS changes in the
enterprise branch as well.

This leaves a problem when it comes to attributing test failures for OSS
tests: The code ownership would be centralized to a very small team if
we keep the same pattern. However, core-structure can handle the load of
guiding good OSS contribution patterns, but cannot handle the load of
triaging arbitrary failures across the tree.

This commit introduces an alternative approach: Define a separate set of
test owners for relevant test code paths in the tree, using a mapping of
OSS codeowner teams to internal teams.

As the new TESTOWNERS.enterprise file is specified last in the parameter
list when passed to the tooling that interprets test results, these
owners will override the owners from CODEOWNERS for the specified paths.

Signed-off-by: Joe Stringer <joe@isovalent.com>