gha: fix overlapping PodCIDR E2E test broken assertion
Currently, the inter-cluster communication test is broken (in other
words it is always green), because the default shell (if not explicitly
specified) is `bash -e {0}`. Given that `-o pipefail` is not set, the
failure of a command piped into another is simply ignored, and does not
cause the interruption of the test.

Additionally, the variables in the check if are not quoted, causing an
error during the final comparison, which oddly is not treated as a hard
failure again:

  line 12: [: !=: unary operator expected

In other words, the whole test completes successfully, even though it
should have failed. Let's get this fixed, by explicitly configuring the
bash shell (which enables `-o pipefail` as well [1]) and quoting the
variables in the check. While being there, let's also configure a
timeout for the curl request, to prevent it from hanging for long
periods of time in case of problems.

[1]: https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#defaultsrunshell

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>