fqdnha: agent: remove restored rules functionality
There are two types of FQDN rules: restored rules and current proxy
rules (a.k.a. "allow" rules). Restored rules are a distilled form of the
allow rules, cached on disk so that the agent can resume serving DNS
very early on startup. Once the agent has started up, the restored rules
are removed and replaced with the up-to-date allow rules derived from policy.

This was once relevant for the non-mindful fqdn-ha proxy, as it was
started very early. That is no longer the case; the doubleproxy is
injected after all endpoints are regenerated, so restored rules are, by
definition, no longer useful.

The FQDN-proxy persists this awkwardness -- we dump all rules in the
"restored" format, then immediately remove these restored rules when the
agent reconnects to the proxy. So, stop this entirely; the agent will
supply the set of rules quickly to the fqdn-proxy, so there's no point
in dealing with the restored rules at all.

For backwards compatability, this continues to call RemoveRestoredRules
from the agent to the remote proxy whenever UpdateAllowed is called.
This preserves the existing behavior as far as an older remote proxy is
concerned.

This also removes other unused / stubbed out methods.

Signed-off-by: Casey Callendrello <cdc@isovalent.com>