fqdnha: proxy: remove RestoreRules functionality
Prevously, the fqdnha-proxy would act similar to the agent proxy, in
that it supported the dual configuration sources RestoreRules() as well
as UpdateAllowed(). However, since the change to the mindful proxy, all
restored rules are immediately replaced by updated rules.

So, don't bother retrieving the set of restored rules from the agent.
The agent will (quickly) connect to us and send the full set of updates
anyways.

In order to prevent any spurious UnimplementedErrors from pre-v1.17
agents, this preserves the proxy-side gRPC method but makes it a noop.

Signed-off-by: Casey Callendrello <cdc@isovalent.com>