clustermesh: workaround skip tunnel flag breaking overlapping PodCIDR
e8db5d317788 ("Fix NodePort traffic to host netns backends with Geneve+DSR")
started setting the skip tunnel flag (previously thought for mixed routing
mode only, which is not compatible with overlapping PodCIDR) for node
entries. This turned out breaking overlapping PodCIDR support; as a short
term workaround, let's always reset the flag (and the tunnel endpoint) when
this feature is enabled. The original change is meant to fix an issue with
DSR+Geneve, so ignoring that looks a reasonable trade-off to make basic
overlapping PodCIDR support work again.

Reported-by: Julian Wiedmann <jwi@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>