fqdnha: remoteproxy: switch to ProxyConfig table
When the remote fqdn proxy is pre-1.17, it opens a gRPC server and
expects a stream of UpdateAllowed() calls. This is implemented by the
RemoteProxy.

Prior to this change, RemoteProxy would shim through DoubleProxy and
intercept all UpdateAllowed calls, adding them to a queue. A separate
goroutine would consume the queue and forward to the remote proxy.

Now, RemoteProxy consumes the table populated by the DoubleProxy,
forwarding all existing and new rows to the remote proxy.

Signed-off-by: Casey Callendrello <cdc@isovalent.com>