cilium-cli: Move CurlCommandParallelWithOutput to enterprise
This doesn't have to be enterprise, but it's currently only used by
internal testsuites. Move the function over to a file pattern that
matches `*enterprise*` so it flies under the radar for regular
'core-structure' pattern checks.

We could upstream this in future if we need the function in an upstream
test, but we don't want to reintroduce it unused upstream (as then
someone will likely come back and remove it again; this happened
previously)

    $ git log --oneline --grep CurlCommandParallelWithOutput
    5ca1b39b5696 enterprise/hooks/connectivity/tests: Revert test changes
    66116d03fb51 Revert "cilium-cli: remove unused CurlCommandParallelWithOutput"
    86e5e23328cc cilium-cli: remove unused CurlCommandParallelWithOutput
    dfd23b97ec79 check: add CurlCommandParallelWithOutput helper

Signed-off-by: Joe Stringer <joe@isovalent.com>