ilb/status: Fix BGP Peers status reporting
Previously, the BGP Peers status reporting in "cilium-cli lb status"
assumed, that a BGP peer name must equal to a corresponding LBService
name. This is obviously wrong.

Before diving into details, it's worth to clarify the meaning of the
"BGP Peers" status. The status field should report a health of BGP
sessions between T1 nodes and BGP peers which are "interested" in
receiving BGP routes for a given LBService.

The commit fixes the issue by:

1. Deriving BGP Peers from BGP ClusterConfig which selects T1 nodes.
2. Deriving a list of BGP Peers for a given LBService. This is done by
   finding IsovalentBGPAdvertisements for the LBService, and then
   finding IsovalentBGPPeerConfigs which match the advertisements.
3. Filtering out the BGP peers from (2) by using (1).

Reported-by: Michael Kashin <michael.kashin@isovalent.com>
Signed-off-by: Martynas Pumputis <martynas@isovalent.com>