hubble-timescape: Avoid unnecesary search domain lookup for vector
We switch to using a FQDN with a dot (.) to avoid search domain lookups.
However, this comes with some SNI shenanigans. Vector, like for example
the majority of browsers, adds the trailing dot to the SNI name. Go, on
the other hand, refuses any SNI name with a trailing dot, as that is
technically invalid according to the RFC.

We can manually set the SNI name for vector, but we still need to update
the certificates to support both names as vector will validate the
certificate based on the FQDN in the URI.

Signed-off-by: Fabian Fischer <fabian.fischer@isovalent.com>