egwha: add PolicyConfig table and store policy state.
Currently policyConfig state is stored in a map, keyed
on the policyID of the policy state we're storing.

We can change this to use a statedb, based on the *PolicyConfig
type:

* Primary index: is the policyID (namespaced-name).

This changes all upserts/deletes/lookups into policyConfig
into equivalent statedb operations.

There are three places where we read/write to the table:

* K8s resource updates: Updates/lookups are performed atomically.
* Reconciliation: We open a writeTxn for the whole reconcileLocked(...)
	function and commit it afterwards.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>