ilb: use snat & maglev for UDPProxy with deployment mode t1-only
Relying on CT map for UDP "session" stickiness doesn't work if
a different T1 node is selected by the "client" in front of the ILB,
because the CT state isn't shared between T1 nodes.

Therefore, this commit is configuring snat & maglev for `LBService`s
with application type `UDPProxy`. This way, UDP packets with the
same source IP & port are sent to the same backend no matter what
T1 node handles the request.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>