ilb: define ciliumnodes & lbdeployments as input testdata
This commit refactors the ILB LBService reconciliation so that
the `ingestor` expects all CiliumNodes and LBDeployments as input.
The ingestor then extracts the relevant data and puts them into the
internal model (T1 & T2 node ips & label selectors).

This way it becomes possible to define the ciliumnodes & lbdeployments
as input testdata for the ILB translation unit tests instead of
having these data hardcoded for all tests.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>