ci: override image without causing backport issue
In this commit we adopt a mechanism same as set-env-variable for set-runtime-image, instead of overriding runtime-image.txt,
we create another txt with the wolfi versions of cilium runtime in it, and we override it on the action itself. The action should not move that much and will provoke
way less backporting issue. We also update the update script to let it point to the new file, enterprise-runtime-image.txt.

Signed-off-by: Antony Reynaud <antony.reynaud@isovalent.com>