egwha: distinguish between agent and operator policy config.
PolicyConfig is used by both agent and operator model and
store current policy state.
However, some fields in PolicyConfig are only used by agent
manager, such as local gateway configs.

Instead, we will now use a compositional type: AgentPolicyConfig
that is a superset of the base PolicyConfig.

This acts the same as policy-configs except has fields
such as local gateway configs.

With this we can store policy-config data in state
db tables we want to be able to differentiate the two
different types a bit.

As well, this allows us to seperate possible implementations
details on things such as statedb table/header printer.

Agent manager control plane should be unaffected, outside of
the type change.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>