envoy: Set socket linger config on HTTP listeners
[ upstream commit 82455fce42eb6061e1dac5a6f1daa8ba3874cb3f ]

Set socket linger config to 0 seconds, to be applied to upstream Envoy
connections for HTTP filter chains that also use the original source
address and port. This should help avoid bind errors on new upstream
connection when an earlier socket with the same source address and port
is still lingering.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>