gha: test host firewall enabled/disabled in mixed routing mode workflow
Originally, host firewall was enabled in all matrix entries of the mixed
routing mode workflow, to better match the expected customer configuration.
However, host firewall triggers a different behavior around pod-to-remote
traffic (which goes through the tunnel to preserve the source identity)
when Cilium is configured in encapsulation mode. This, in turn, led to
missing a regression when host firewall is disabled. Hence, let's make
sure to cover both cases as part of this test.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>