(cee) bpf: encrypt: avoid looking up source identity
When we have a valid security identity for the source endpoint, avoid the
ipcache lookup. This prevents leaks if the source endpoint has been torn
down by the time that its packets reach the to-netdev program, and the
ipcache lookup falls through to the catch-all entry (which is WORLD_ID,
and thus not eligible for encryption).

In this patch we either use the source identity that's extracted by
the to-netdev program, or (for overlay traffic) the identity that's
embedded into the skb->mark.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>