(cee) bpf: tests: set accurate identity mark for encryption tests
A skb->mark with MARK_MAGIC_OVERLAY also contains the security identity
of the source endpoint. Fix this up, so that the test reflects the actual
datapath behavior.

In the same way, set the MARK_MAGIC_IDENTITY mark for packets in the
routing case. This would usually be set by the from-container program.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>