cilium-cli:feature: add TunnelPort to CLI features
[ backporter's notes: fixed enterprise EncryptionPolicy and MixedRouting ]

This commit adds TunnePort to the set of features detected from our CLI
during tests. This can be useful from tests to retrieve the current
active tunnel port. The `Enabled` is set to false in case `tunnel-port`
is not present in the ConfigMap, but its value `Mode` is always set to
either the provided port or the default one for the currently active
protocol.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>