cilium-cli:fix: use tunnel-port in use when building TunnelFilter
[ backporter's notes: synced change with MixedRouting and EncryptPolicy.
                      Added a new enterprise-only GetTunnelFilter() function
                      that leverages the OSS one to retrieve tunnelFilter
                      of the provided configuration. Retrieve both local
                      and remote tunnelFilter and use them both to address
                      the MixedRouting use case. ]

This commit adjusts the current TunnelFilter logic to account for the
current tunnel-port in use when building tcpdump filters for the overlay.
This should prevent CLI tests from failing in all those environments where
tunnel-port is not the default one (ex. 8472 VXLAN, 6081 Geneve).

Given that now the filter `(udp and port tunnelPort)` make use of the
current tunnel port in use, we don't need anymore additional hacks:

- VXLAN: "udp[8:2] = 0x0800" would compare the first two bytes of an UDP
    payload against VXLAN commonly used flags.
- Geneve: we could not use the "geneve" filter, as it shifts offset of a
    filtered packet, which invalidates a filter matching on the outer headers.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>