hubble-timescape: Update default aggregation to not renew the ttl
We were a bit too aggressive with the default aggregation. Without this,
we can miss long standing active connections, as we'll only export the
flow once. Then if we see a flow for that connection at least every
30s, we don't export the flow again as the TTL will be renewed every
time.

This change disables the TTL renew. So we'll emit one flow every 30s for
active connections.

Signed-off-by: Fabian Fischer <fabian.fischer@isovalent.com>