renovate: Group hubble images update together
We currently don't group any of the image dependencies in for the helm
chart. This leads to a bunch of very similar renovate requests,
especially when we make a new hubble timscape release as 3 images are
released at the same time.

Example:

* https://github.com/isovalent/cilium/pull/6697
* https://github.com/isovalent/cilium/pull/6804
* https://github.com/isovalent/cilium/pull/6809

This is made worse by the fact, that changes in the values file won't
request review from hubble, which are the owners of these dependencies.
Let's explicitly request review from them.

Signed-off-by: Fabian Fischer <fabian.fischer@isovalent.com>